+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * fixed bug on the backend topbar. The topbar is entirely shown only on mobile devices. Added the sidebar;
 * widgets are hidden on the pages that contain the same information or the same data;
 * checks if the latest search has been executed out of the minimum interval;
+* shows the version of Apache and PHP;
 * KCFinder permissions are based on MeCms users permissions;
 * many buttons are disabled after the click, to prevent some actions are performed repeatedly;
 * usernames and the name of user groups cannot be changed. Improved permissions about users and user groups;